Review Comment fixes
